requirements.txt
Flask==2.0.1
Flask_SQLAlchemy==2.5.1
python-dotenv==0.17.1
